fix output paths for imports that are relative to a common ancestor; move document writing to separate function shared by Fix, Format, and Upgrade
