+++ src/main/scala/wdlTools/cli/Fix.scala
@@ -1,13 +1,12 @@
 package wdlTools.cli
 
-import java.nio.file.{Path, Paths}
+import java.nio.file.Paths
 import wdlTools.generators.code.WdlGenerator
-import dx.util.{AddressableFileSource, FileNode, FileSourceResolver, FileUtils, LinesFileNode}
+import dx.util.{AddressableFileSource, FileNode, FileSourceResolver, LinesFileNode}
 import wdlTools.syntax.{Parsers, SyntaxException}
 import wdlTools.types.TypedAbstractSyntax.{Document, ImportDoc}
-import wdlTools.types.{TypeCheckingRegime, TypeException, TypeInfer}
+import wdlTools.types.TypeCheckingRegime
 
-import java.io.{ByteArrayOutputStream, PrintStream}
 import scala.collection.immutable.SeqMap
 import scala.language.reflectiveCalls
 
@@ -15,50 +14,12 @@ case class Fix(conf: WdlToolsConf) extends Command {
   override def apply(): Unit = {
     val fileResolver = FileSourceResolver.get
     val docSource = fileResolver.resolve(conf.fix.uri())
-    val baseFileSource =
-      conf.fix.baseUri.toOption
-        .map(fileResolver.resolveDirectory(_))
-        .orElse(docSource.getParent)
-        .getOrElse(
-            throw new Exception(s"cannot determine the base FileSource for main WDL ${docSource}")
-        )
     val srcVersion = conf.fix.srcVersion.toOption
-    val outputDir = conf.fix.outputDir.toOption.getOrElse(Paths.get("."))
-    val overwrite = conf.fix.overwrite()
     val wdlVersion = srcVersion.getOrElse(Parsers.default.getWdlVersion(docSource))
 
-    def parseAndCheck(
-        source: FileNode,
-        message: String,
-        regime: TypeCheckingRegime.TypeCheckingRegime,
-        fileResolver: FileSourceResolver
-    ): Document = {
-      val errorHandler = new ErrorHandler()
-      val parsers = Parsers(conf.fix.followImports(), fileResolver = fileResolver)
-      val parser = parsers.getParser(wdlVersion)
-      val checker =
-        TypeInfer(regime, fileResolver = fileResolver, errorHandler = Some(errorHandler.apply))
-      val document =
-        try {
-          val (document, _) = checker.apply(parser.parseDocument(source))
-          document
-        } catch {
-          case e: SyntaxException =>
-            throw new Exception(s"Failed to parse ${source}; ${message}", e)
-          case e: TypeException =>
-            throw new Exception(s"Failed to type-check ${source}; ${message}", e)
-        }
-      if (errorHandler.hasErrors) {
-        val msgStream = new ByteArrayOutputStream()
-        errorHandler.printErrors(new PrintStream(msgStream), effects = true)
-        throw new Exception(
-            s"Failed to type-check ${source}; ${message}:\n${msgStream.toString()}"
-        )
-      }
-      document
-    }
-
     val document = parseAndCheck(docSource,
+                                 wdlVersion,
+                                 conf.fix.followImports(),
                                  "this error cannot be fixed automatically",
                                  TypeCheckingRegime.Lenient,
                                  fileResolver)
@@ -66,16 +27,13 @@ case class Fix(conf: WdlToolsConf) extends Command {
     // use a SeqMap so that the first entry will be the file generated from the source document
     val generator =
       WdlGenerator(Some(wdlVersion), omitNullCallInputs = false, rewriteNonstandardUsages = true)
-    var results: SeqMap[Path, Iterable[String]] = SeqMap.empty
-    var visited: Set[String] = Set.empty
+    var results: SeqMap[AddressableFileSource, Iterable[String]] = SeqMap.empty
 
     def generateDocument(fileSource: FileNode, doc: Document): Unit = {
       fileSource match {
-        case fs: AddressableFileSource if !visited.contains(fileSource.toString) =>
-          visited += fileSource.toString
-          val outputPath = outputDir.resolve(baseFileSource.relativize(fs))
+        case fs: AddressableFileSource if !results.contains(fs) =>
           val lines = generator.generateDocument(doc)
-          results += outputPath -> lines
+          results += fs -> lines
           doc.elements.foreach {
             case ImportDoc(_, _, addr, doc) =>
               generateDocument(fileResolver.resolve(addr, Some(fs)), doc)
@@ -91,9 +49,9 @@ case class Fix(conf: WdlToolsConf) extends Command {
     // validate that the generated files parse
     val validationParser = Parsers.default.getParser(wdlVersion)
     results.foreach {
-      case (path, lines) =>
+      case (fs, lines) =>
         try {
-          validationParser.parseDocument(LinesFileNode(lines.toVector, path.toString))
+          validationParser.parseDocument(LinesFileNode(lines.toVector, fs.toString))
         } catch {
           case e: SyntaxException =>
             throw new Exception(
@@ -103,20 +61,19 @@ case class Fix(conf: WdlToolsConf) extends Command {
         }
     }
 
-    // write out fixed versions
-    results.foreach {
-      case (path, lines) =>
-        FileUtils.writeFileContent(path, lines.mkString(System.lineSeparator()), overwrite)
-    }
-
     // Validate that the fixed versions type-check correctly - we can't do this until
     // after we write out the documents because otherwise import resolution won't work -
-    // this should only fail if there is a bug in the code generator. Prepend the output
-    // dir to the search path.
+    // this should only fail if there is a bug in the code generator.
+    val outputDir = conf.fix.outputDir.toOption.getOrElse(Paths.get("."))
+    val overwrite = conf.fix.overwrite()
+    val outputPaths = writeDocuments(results, Some(outputDir), overwrite)
+    // Prepend the output dir to the search path.
     val rewrittenFileResolver = fileResolver.addToLocalSearchPath(Vector(outputDir), append = false)
     parseAndCheck(
-        rewrittenFileResolver.fromPath(results.head._1),
-        "this is likely due to a bug in the WDL code generator - please report this issue",
+        rewrittenFileResolver.fromPath(outputPaths(results.head._1)),
+        wdlVersion,
+        conf.fix.followImports(),
+        "this may be an unsupported fix, or it may be due to a bug in the WDL code generator - please report this issue",
         TypeCheckingRegime.Moderate,
         rewrittenFileResolver
     )
